pyproject.toml
[project]
name = "rag-project"
version = "0.1.0"
description = "A modular RAG → fine‑tune pipeline."
authors = ["Your Name <you@example.com>"]
license = "MIT"
dependencies = [
    "pydantic>=1.10",
    "faiss-cpu>=1.7.4",
    "openai>=1.0.0",
    "sentence-transformers>=2.2.2",
    "numpy>=1.24",
    "fastapi>=0.110",
    "uvicorn>=0.23",
    "python-dotenv>=1.0",
]
requires-python = ">=3.9"

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "ruff>=0.1",
    "black>=23.7",
]
vector_backends = [
    "chromadb>=0.4.13",
    "pinecone-client>=2.2.4",
]
cli = [
    "typer>=0.9",
]

[tool.black]
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "W"]


[tool.pytest.ini_options]
pythonpath = ["."]
